Add missing query tests
