--- mutation/create-user/index.ts
@@ -1,7 +1,0 @@
-import { describe } from 'mocha';
-
-describe('createUser', () => {
-  require('./user-creation');
-  require('./user-input-validation');
-  require('./authentication');
-});
