--- mutation/index.ts
@@ -1,6 +1,0 @@
-import { describe } from 'mocha';
-
-describe('mutation', () => {
-  require('./create-user');
-  require('./login');
-});
