--- query/index.ts
@@ -1,5 +1,0 @@
-import { describe } from 'mocha';
-
-describe('query', () => {
-  require('./user');
-});
